Finished remaking the request form
I had to make a new model, form, update the view, and updated the html for the form. Now when a form is submitted, it is sent to the database. It can later be sent to an email as well. I also added a success page for when a form submitts successfully.